Finished testing PlayerDAOIMPIT

+++ com.boracompany.mygame/src/main/java/com/boracompany/mygame/ORM/PlayerDAO.java
@@ -6,7 +6,7 @@ import com.boracompany.mygame.Model.Player;
 
 public interface PlayerDAO {
 	   public List<Player> getAllPlayers();
-	   public Player getPlayer(int ID);
+	   public Player getPlayer(Long ID);
 	   public void updatePlayer(Player player);
 	   public void deletePlayer(Player player);
 }
